+++ src/polymorphic/tests/admin.py
@@ -1,6 +1,33 @@
-from django.contrib.admin import register, ModelAdmin
-from polymorphic.admin import StackedPolymorphicInline, PolymorphicInlineSupportMixin
-from polymorphic.tests.models import PlainA, InlineModelA, InlineModelB, InlineParent
+from django.contrib.admin import register, ModelAdmin, site as admin_site
+from polymorphic.admin import (
+    StackedPolymorphicInline,
+    PolymorphicInlineSupportMixin,
+    PolymorphicChildModelAdmin,
+    PolymorphicChildModelFilter,
+    PolymorphicParentModelAdmin,
+)
+
+from polymorphic.tests.models import (
+    PlainA,
+    Model2A,
+    Model2B,
+    Model2C,
+    Model2D,
+    InlineModelA,
+    InlineModelB,
+    InlineParent,
+)
+
+
+@register(Model2A)
+class Model2Admin(PolymorphicParentModelAdmin):
+    list_filter = (PolymorphicChildModelFilter,)
+    child_models = (Model2B, Model2C, Model2D)
+
+
+admin_site.register(Model2B, PolymorphicChildModelAdmin)
+admin_site.register(Model2C, PolymorphicChildModelAdmin)
+admin_site.register(Model2D, PolymorphicChildModelAdmin)
 
 
 @register(PlainA)