+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 v4.2.0 (2025-12-01)
 -------------------
 
+* Fixed `Show full admin context (breadcrumb and logout nav) in model type selection admin form <https://github.com/jazzband/django-polymorphic/pull/580>`_
 * Fixed `Issue with Autocomplete Fields in StackedPolymorphicInline.Child Inline <https://github.com/jazzband/django-polymorphic/issues/546>`_
 * Support Python 3.14 and Django 6.0, drop support for EOL python 3.9, Django 3.2, 4.0, 4.1 and 5.0.
 * `Modernized package management with new build, test, docs tooling and improved CI